Email-Workflow-Automation-main/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4~=4.12.3
requests
pydantic~=2.9.2
google-generativeai
langchain
clickup